celery/requirements.txt
celery[redis]==5.3.6
python-dotenv==1.0.0